% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds_functions_public.R
\name{matrix_makarov}
\alias{matrix_makarov}
\title{Matrix Makarov Bounds for Distribution of Treatment Effects at a Point}
\usage{
matrix_makarov(
  Treat.Mat,
  Control.Mat,
  y,
  smooth = FALSE,
  bandwidth = NULL,
  n.grid = 20,
  hc = FALSE
)
}
\arguments{
\item{Treat.Mat}{Matrix of outcomes for treated units}

\item{Control.Mat}{Matrix of outcomes for control units}

\item{y}{Point at which to bound DTE}

\item{smooth}{Logical indicating whether the smoothing should be performed. Smoothing is recommended if potential outcomes are estimated.}

\item{bandwidth}{Bandwidth to be used for smoothing}

\item{n.grid}{Number of grid points to search over for evaluating the bounds at each point.}

\item{hc}{Logical indicating whether or not to correct for row and column heterogeneity. See Section 5.2 of Auerbach and Cai (2023).}
}
\value{
A list of bounds and the components.\tabular{ll}{
\code{bounds} \tab Lower and upper Makarov bounds for the DTE at \eqn{y} \cr
\tab \cr
\code{breakdown} \tab Breakdown of the bounds into the vectorized Frechet-Hoeffding (vf) and eigenvalue-based (evb) components. Trivial bounds for probability are 0 and 1. \cr
}
}
\description{
This function computes the bounds for the distribution of treatment effects (DTE) at a given point. Specifically, it computes bounds for \eqn{\Delta(y)} for a specified \eqn{y}
}
\examples{
# Generate data
Treat.Mat <- matrix(runif(200) > 0.8, 10, 20)*1
Control.Mat <- matrix(runif(100) > 0.8, 10, 10)*1
  
# Generate DTE and plot it
bounds <- matrix_makarov(Treat.Mat, Control.Mat, 0.5)
}
