% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixHTE-package.R
\docType{package}
\name{MatrixHTE-package}
\alias{MatrixHTE}
\alias{MatrixHTE-package}
\title{MatrixHTE: Heterogeneous Treatment Effects for Networks, Panels and Other Outcome Matrices}
\description{
This package contains functions for implementing the methods proposed in Identifying Social Disruptions (Auerbach and Cai, 2023). For a complete list of functions, use \code{library(help = "MatrixHTE")}.
}
\author{
\strong{Maintainer}: Yong Cai \email{yongcai@uchicago.edu}

Authors:
\itemize{
  \item Eric Auerbach \email{eric.auerbach@northwestern.edu}
}

}
\keyword{internal}
