% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds_functions_public.R
\name{plot.dte}
\alias{plot.dte}
\title{Plot Distribution of Treatment Effects}
\usage{
\method{plot}{dte}(dte_obj)
}
\arguments{
\item{dte_obj}{Object of class \code{dte} computed by \code{dte_bounds()}.}
}
\value{
A \code{ggplot2} plot of the DTE.
}
\description{
Method for plotting DTE.
}
\examples{
# Generate data
Treat.Mat <- matrix(runif(200) > 0.8, 10, 20)*1
Control.Mat <- matrix(runif(100) > 0.8, 10, 10)*1

# Generate DTE and plot it
bounds <- dte_bounds(Treat.Mat, Control.Mat, -5, 5, smooth = TRUE)
plot(bounds)
}
