% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds_functions_public.R
\name{ste}
\alias{ste}
\title{Spectral Treatment Effects}
\usage{
ste(Treat.Mat, Control.Mat, hc = FALSE)
}
\arguments{
\item{Treat.Mat}{Matrix of outcomes for treated units}

\item{Control.Mat}{Matrix of outcomes for control units}

\item{hc}{Logical indicating whether or not to incorporate row and column heterogeneity. See Section 5.2 of Auerbach and Cai (2022).}
}
\value{
An object of class \code{ste} containing \tabular{ll}{
\code{stt} \tab STE on treated units sorted in increasing order.\cr
\tab \cr
\code{stc} \tab STE on control units sorted in increasing order.\cr
\tab \cr
\code{treat.cf} \tab Counterfactual for treated units based on eigenvalue imputation. \cr
\tab \cr
\code{control.cf} \tab Counterfactual for control units based on eigenvalue imputation. \cr
}
}
\description{
This function computes spectral treatment effects as described in Section 3.2.2 of Auerbach and Cai (2022).
}
\examples{
# Generate Data
temp <- matrix(rnorm(100), 10,10)
temp <- (temp + t(temp))/2
Treat.Mat <- temp*1
diag(Treat.Mat) <- 0

temp <- matrix(rnorm(400), 20,20)
temp <- (temp + t(temp))/2
Control.Mat <- temp*1
diag(Control.Mat) <- 0
# Compute STE
effects <- ste(Treat.Mat, Control.Mat)
# Compute CDF of STE and plot them
effects_cdf <- ste_cdf(effects)
plot(effects_cdf$stt)
plot(effects_cdf$stc)
}
