% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds_functions_public.R
\name{svt.smooth}
\alias{svt.smooth}
\title{Smooth Matrix via SVT}
\usage{
svt.smooth(input.mat, p1, bandwidth)
}
\arguments{
\item{input.mat}{Matrix to be smoothed}

\item{p1}{Density of the matrix}

\item{bandwidth}{Bandwidth to be used for smoothing. Default is eta = 0.01 as recommended in Chatterjee (2015).}
}
\value{
Smoothed input
}
\description{
This function smooths a matrix using singular vector thresholding as in Chatterjee (2015).
}
