% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds_functions_public.R
\name{dpo_bounds}
\alias{dpo_bounds}
\title{Bounds for Distribution of Potential Outcomes}
\usage{
dpo_bounds(
  Treat.Mat,
  Control.Mat,
  y1,
  y0,
  smooth = FALSE,
  bandwidth = NULL,
  hc = FALSE
)
}
\arguments{
\item{Treat.Mat}{Matrix of outcomes for treated units}

\item{Control.Mat}{Matrix of outcomes for control units}

\item{y1}{y1 in F(y1, y0).}

\item{y0}{y0 in F(y1, y0).}

\item{smooth}{Logical indicating whether the smoothing should be performed. Smoothing is recommended if potential outcomes are estimated.}

\item{bandwidth}{Bandwidth to be used for smoothing}

\item{hc}{Logical indicating whether or not to incorporate row and column heterogeneity. See Section 5.2 of Auerbach and Cai (2022).}
}
\value{
Lower and upper bounds for the DPO.
}
\description{
This function implements the bounds for the distribution of potential outcomes as described in Proposition 1 of Auerbach and Cai (2022).
}
\examples{
# Generate data
Treat.Mat <- matrix(runif(200) > 0.8, 10, 20)*1
Control.Mat <- matrix(runif(100) > 0.8, 10, 10)*1

# P(Treat.Mat <= 0.5, Control.Mat <= 0.5)
dpo_bounds(Treat.Mat, Control.Mat, 0.5, 0.5)

# P(Treat.Mat == 1, Control.Mat == 1)
dpo_bounds(-Treat.Mat, -Control.Mat, -0.5, -0.5) # input matrices are binary
}
