% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounds_functions_public.R
\name{dte_bounds}
\alias{dte_bounds}
\title{Bounds for Distribution of Treatment Effects}
\usage{
dte_bounds(
  Treat.Mat,
  Control.Mat,
  start,
  end,
  dte.n.grid = 30,
  smooth = FALSE,
  bandwidth = NULL,
  n.grid = 20,
  hc = FALSE,
  eval.points = NULL
)
}
\arguments{
\item{Treat.Mat}{Matrix of outcomes for treated units}

\item{Control.Mat}{Matrix of outcomes for control units}

\item{start}{Starting value for evaluating the DTE}

\item{end}{End value for evaluating the DTE}

\item{dte.n.grid}{Number of points between start and end at which the CDF is bounded.}

\item{smooth}{Logical indicating whether the smoothing should be performed. Smoothing is recommended if potential outcomes are estimated.}

\item{bandwidth}{Bandwidth to be used for smoothing}

\item{n.grid}{Number of grid points to search over for evaluating the bounds at each point.}

\item{hc}{Logical indicating whether or not to incorporate row and column heterogeneity. See Section 5.2 of Auerbach and Cai (2022).}

\item{eval.points}{Vector indicating additional points at which DTE should be explicitly evaluated.}
}
\value{
Lower and upper bounds for the DTE over the grid points.
}
\description{
This function computes the bounds for the distribution of treatment effects (DTE) along a sequence of points. Specifically, it computes bounds for \eqn{\Delta(y)} for \eqn{y \in [start, end]}. The output of this function can be plotted directly using \code{plot()}.
}
\examples{
# Generate data
Treat.Mat <- matrix(runif(200) > 0.8, 10, 20)*1
Control.Mat <- matrix(runif(100) > 0.8, 10, 10)*1

# Generate DTE and plot it
bounds <- dte_bounds(Treat.Mat, Control.Mat, -5, 5, smooth = TRUE)
plot(bounds)
}
